Aoc7 finished

+++ src/main/java/cz/mawa/aoc13/IntegerValue.java
@@ -8,7 +8,7 @@ public class IntegerValue extends TreeNode {
         this.value = value;
     }
 
-    public Long getValue() {
+    public Integer getValue() {
         return value;
     }
